GV44: re-record remaining fixture replays, and stop the set going stale
CI's wasm-replay-viewer job rejected two fixtures the first GV44 pass missed:
gen-small-pits and gen-colossal-4team were still GV43-stamped. Root cause is
structural, not carelessness — NO native test opens either file. They exist
solely for the wasm smoke (a small GENERATED map, because generated maps dig
trenches and the hand-authored arena digs none; and a 4992x4992 board as the
wasm32 address-space canary), so a re-record pass driven by the test files
finds four of six and the shards stay green.

Full inventory swept by header stamp, not by guesswork:

  capture-seed1        GV44 already  teams=2  (first pass)
  wipe-lives1          GV44 already  teams=2  (first pass)
  draw-nokill          GV44 already  teams=2  (first pass)
  tests/replays/ctf    GV44 already  teams=2  (first pass)
  gen-small-pits       GV43 -> GV44  teams=2  re-recorded
  gen-colossal-4team   GV43 -> GV44  teams=4  re-recorded

Neither gen-* recipe was written down anywhere (colossal's lived only in a
build.yml comment; small-pits' nowhere at all), so both were recovered from
the fixtures themselves — a replay embeds the full config JSON it was
recorded with. Same seeds and knobs as the originals, on a server rebuilt
from this tree:
  record_fixture.sh gen-small-pits 4242 1500
    '{"mapPath":"gen","mapSeed":4242,"mapSize":"small"}'
  record_colossal_demo.sh gen-colossal-4team 4242 1500 16

PORTABILITY FIX, found while verifying the above. `homeRotationFor` mixed
its inputs with `cast[uint64](value)`. `int` is 64 bits on the server and 32
bits in the wasm viewer, and a WIDENING cast leaves the high half undefined —
and the viewer re-derives the deal when it re-simulates. That is a silent
server/viewer desync on every 4-team replay in the observatory, invisible to
the native suite by construction. Now mixed as `uint64(cast[uint32](value))`,
a same-size reinterpretation that is well-defined and identical on both
targets. gen-colossal-4team was re-recorded AFTER this fix, since it changes
the deal.

GUARD, so this cannot recur: test_replay gains "EVERY committed .bitreplay
carries the current GameVersion", which walks tests/ rather than the test's
own reading list and loads each file through the codec's own gate. Verified
to DISCRIMINATE, not just pass — restoring the GV43 gen-small-pits makes it
fail and name the file plus the recipe pointer. AGENTS.md now carries the
full six-fixture table with recipes and the "the shards only read four"
warning.

VERIFICATION. No local emsdk/Docker bundle, so the wasm smoke's native
equivalent was run instead: every fixture loaded through the same
replay_runtime the viewer wraps and stepped 300 frames, demanding
hashMismatchTick == -1 throughout. All six clean — including
gen-colossal-4team, which now carries a live rotation (rot=3) and
re-simulates exactly, which is the end-to-end proof that the deal survives
the replay path.

Full suite green: shard_1 160, shard_2 165, shard_3 120, shard_4 195 = 640
checks, 0 failures; nim check src/ctf.nim clean.

+++ src/ctf/arena.nim
@@ -445,7 +445,14 @@ proc homeRotationFor*(seed, teams: int): int =
     return 0
   var hash = 14695981039346656037'u64
   for value in [seed, teams]:
-    hash = hash xor cast[uint64](value)
+    ## Mixed as the low 32 bits of the two's-complement pattern, NOT as
+    ## `cast[uint64](value)`: `int` is 64 bits on the server and 32 bits in
+    ## the wasm replay viewer, and a widening `cast` leaves the high half
+    ## undefined. The viewer re-derives this deal when it re-simulates a
+    ## replay, so a word-size disagreement here would desync every 4-team
+    ## replay in the observatory while the native suite stayed green.
+    ## Same-size `cast` to uint32 is a well-defined reinterpretation on both.
+    hash = hash xor uint64(cast[uint32](value))
     hash = hash * 1099511628211'u64
   hash = hash xor (hash shr 30)
   hash = hash * 13787848793156543929'u64
